+++ src/changelog.md
@@ -1,12 +1,21 @@
 # Manifold Pumps Changelog
 
-# # Version 0.0.1
+## Version 0.0.1
 - Initial release for 0.16
 
-# # Version 0.0.2
+## Version 0.0.2
 - Updates to work on 0.17
 - Added thumbnail
 
-# # Version 0.0.3
+## Version 0.0.3
 - Increase fluid box height as this allows even more insane speeds (pump now buffers 1k fluid)
 - Add icons that show pump direction
+
+## Version 0.0.5
+- Update for Factorio v1.1
+
+## Version 0.0.6
+- Update for Factorio v2.0 and Space Age!
+- Require tungsten plate when Space Age is enabled.
+- Remove conversion recipe between manifold directions. There are now just 2 recipes.
+- Increase energy consumption.